move parseLayoutFeatures to util module
